logs: refactor out pipe solana logs
